requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
deepface==0.0.93
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
scipy==1.13.1

